issue9: move cli stuff into repo.cli module
and amend references and tests accordingly
