+++ src/notsotuf/utils/__init__.py
@@ -80,46 +80,3 @@ def input_text(
         if optional:
             break
     return answer
-
-
-def get_config_from_user(**kwargs) -> dict:
-    top_level_role_names = ['root', 'targets', 'snapshot', 'timestamp']
-    for key, example, optional in [
-        ('app_name', '', False),
-        ('app_version_attr', ', e.g. my_app.__version__', True),
-        ('repo_dir', '', False),
-        ('keys_dir', '', False),
-    ]:
-        kwargs[key] = input_text(
-            prompt=f'Specify {key}{example}',
-            default=kwargs.get(key),
-            optional=optional,
-        )
-    key_map = kwargs.get('key_map', {})
-    expiration_days = kwargs.get('expiration_days', {})
-    encrypted_keys = []
-    unique_key_names = []
-    for role_name in top_level_role_names:
-        # key_map
-        key_name = input_text(
-            prompt=f'Specify key name for {role_name}',
-            default=key_map.get(role_name, role_name),
-        )
-        key_map[role_name] = key_name
-        # encrypted_keys
-        if key_name not in unique_key_names:
-            unique_key_names.append(key_name)
-            if input_bool(
-                    prompt=f'Encrypt key "{key_name}"?',
-                    default=key_name in kwargs.get('encrypted_keys', [])
-            ):
-                encrypted_keys.append(key_name)
-        # expiration_days
-        expiration_days[role_name] = input_numeric(
-            prompt=f'Specify number of days before {role_name} expires',
-            default=expiration_days.get(role_name, 1),
-        )
-    kwargs['key_map'] = key_map
-    kwargs['expiration_days'] = expiration_days
-    kwargs['encrypted_keys'] = encrypted_keys
-    return kwargs
