+++ tests/test_utils.py
@@ -97,46 +97,3 @@ class InputTests(unittest.TestCase):
                     prompt='', default=None, optional=True
                 )
             )
-
-    def test_get_config_from_user_no_kwargs(self):
-        default = ''
-        yes = 'y'
-        no = 'n'
-        user_inputs = iter(
-            [
-                'my-app',
-                'my_app.__version__',
-                'repo/dir',
-                'keys/dir',
-                default,
-                yes,
-                '365',
-                default,
-                no,
-                '7',
-                default,
-                no,
-                '7',
-                default,
-                no,
-                '1',
-            ]
-        )
-        with patch('builtins.input', lambda *_, **__: next(user_inputs)):
-            config_kwargs = notsotuf.get_config_from_user()
-        self.assertTrue(config_kwargs)
-
-    def test_get_config_from_user_with_kwargs(self):
-        original_kwargs = dict(
-            app_name='my-app',
-            app_version_attr='my_app.__version__',
-            repo_dir='repo/dir',
-            keys_dir='keys/dir',
-            key_map=notsotuf.repo.DEFAULT_KEY_MAP,
-            encrypted_keys=['root'],
-            expiration_days=notsotuf.repo.DEFAULT_EXPIRATION_DAYS,
-        )
-        default = ''
-        with patch('builtins.input', Mock(return_value=default)):
-            config_kwargs = notsotuf.get_config_from_user(**original_kwargs)
-        self.assertEqual(config_kwargs, original_kwargs)
